move Duration type to model package
