+++ master/internal/config_test.go
@@ -73,8 +73,8 @@ resource_pools:
 						AgentDockerNetwork:     "default",
 						AgentDockerImage:       fmt.Sprintf("determinedai/determined-agent:%s", version.Version),
 						AgentFluentImage:       "fluent/fluent-bit:1.6",
-						MaxIdleAgentPeriod:     provisioner.Duration(30 * time.Second),
-						MaxAgentStartingPeriod: provisioner.Duration(30 * time.Second),
+						MaxIdleAgentPeriod:     model.Duration(30 * time.Second),
+						MaxAgentStartingPeriod: model.Duration(30 * time.Second),
 						MaxInstances:           5,
 					},
 					MaxAuxContainersPerAgent: 100,
@@ -134,8 +134,8 @@ resource_pools:
 						AgentDockerNetwork:     "default",
 						AgentDockerImage:       fmt.Sprintf("determinedai/determined-agent:%s", version.Version),
 						AgentFluentImage:       "fluent/fluent-bit:1.6",
-						MaxIdleAgentPeriod:     provisioner.Duration(10 * time.Second),
-						MaxAgentStartingPeriod: provisioner.Duration(20 * time.Second),
+						MaxIdleAgentPeriod:     model.Duration(10 * time.Second),
+						MaxAgentStartingPeriod: model.Duration(20 * time.Second),
 						MaxInstances:           5,
 					},
 					MaxAuxContainersPerAgent: 10,
@@ -148,8 +148,8 @@ resource_pools:
 						AgentDockerNetwork:     "default",
 						AgentDockerImage:       fmt.Sprintf("determinedai/determined-agent:%s", version.Version),
 						AgentFluentImage:       "fluent/fluent-bit:1.6",
-						MaxIdleAgentPeriod:     provisioner.Duration(30 * time.Second),
-						MaxAgentStartingPeriod: provisioner.Duration(40 * time.Second),
+						MaxIdleAgentPeriod:     model.Duration(30 * time.Second),
+						MaxAgentStartingPeriod: model.Duration(40 * time.Second),
 						MaxInstances:           5,
 					},
 					MaxAuxContainersPerAgent: 0,
